+++ app.py
@@ -1,2 +1,2 @@
-from apiroutes import *
+from routes import *
 print("Starting Thought at \n"+ "http://0.0.0.0:8080 on "+str(bruh.now().time()))
