+++ core.py
@@ -36,10 +36,15 @@ app.config['JWT_EXP_DELTA_SECONDS'] = 3600
 IMAGED = os.environ.get("IMAGES_PATH")
 REPOS_PATH = os.environ.get("REPOS_PATH")
 COMMENT_MAX =2000
+COMMENT_MIN =2
 POST_MAX = 5000
+POST_MIN = 2
 ABOUT_ME_MAX = 50
+ABOUT_ME_MIN = 2
 PASSWORD_MAX = 12
+PASSWORD_MIN = 8
 MESSAGE_MAX = 100
+MESSAGE_MIN = 1
 USERNAME_MAX = 12
 USERNAME_MIN = 2
 STATIC_FOLDER = os.environ.get("STATIC_FOLDER")
