+++ routes.py
@@ -1,4 +1,3 @@
-import os
 from models import *
 from forms import *
 from werkzeug.utils import secure_filename
@@ -6,12 +5,13 @@ from werkzeug.security import generate_password_hash
 import re
 import subprocess
 import shutil
-import base64
 import markdown
-from datetime import datetime
-from datetime import datetime as bruh
-
-from models import Post, Messages, Repository, User
+from datetime import datetime as dt
+import jwt
+from datetime import timedelta
+from functools import wraps
+from core import app, db
+from flask import request, jsonify, g
 
 
 @app.route('/image/posts/<user_id>/<image_name>')
@@ -25,7 +25,7 @@ def image_post(user_id, image_name):
 def format_timestamp(ts):
     if not ts:
         return ""
-    if isinstance(ts, datetime):
+    if isinstance(ts, dt):
         return ts.strftime('%b %d, %Y - %H:%M')
     return str(ts)
 
@@ -83,7 +83,7 @@ def home():
 def user_active():
     db.userdb.update_one(
         {"_id": current_user._id},
-        {"$set": {"last_seen": bruh.now()}}
+        {"$set": {"last_seen": dt.now()}}
     )
     return jsonify({"success": True})
 
@@ -94,9 +94,9 @@ def user_status(username):
     user = User.get_by_username(username)
     if not user:
         return jsonify({"online": False, "last_seen": "Never"}), 404
-    if user.last_seen and isinstance(user.last_seen, datetime):
+    if user.last_seen and isinstance(user.last_seen, dt):
         # Compare current UTC time to last_seen
-        is_online = (datetime.utcnow() - user.last_seen).total_seconds() < 60
+        is_online = (dt.utcnow() - user.last_seen).total_seconds() < 60
         last_seen_str = user.last_seen.strftime("%I:%M %p - %b %d, %Y")
     else:
         is_online = False
@@ -226,7 +226,7 @@ def upload_image():
             flash("No valid image files uploaded!", "danger")
             return redirect("/me")
 
-    return render_template("upload_image.html")
+    return render_template("forms/upload_image.html")
 
 
 @app.route("/image/<userid>/<imageuid>", methods=["GET", "POST"])
@@ -315,7 +315,7 @@ def register():
                     flash('username does not meet requirements')
             else:
                 flash(f'Account already exists for {form.username.data}!', 'success')
-    return render_template('register.html', title='Register', form=form)
+    return render_template('forms/register.html', title='Register', form=form)
 
 
 @app.route("/login", methods=['GET', 'POST'])
@@ -332,13 +332,13 @@ def login():
                 current_user.is_authenticated = True
                 db.userdb.update_one(
                     {"_id": user._id},
-                    {"$set": {"last_seen": bruh.now()}}
+                    {"$set": {"last_seen": dt.now()}}
                 )
                 flash(f'You are now logged in as {form.username.data}!', 'success')
                 return redirect(next or url_for('home'))
             else:
                 flash(f'Invalid login!', 'danger')
-    return render_template('login.html', title='Login', form=form)
+    return render_template('forms/login.html', title='Login', form=form)
 
 
 
@@ -348,12 +348,12 @@ def login():
 @login_required
 def createnewpost():
     form = PostForm()
-
+    p = 0
     if form.validate_on_submit():
         title = form.title.data
         content = form.content.data
         user_id = current_user._id
-        timestamp = bruh.now().strftime('%H:%M:%S %Y-%m-%d')
+        timestamp = dt.now().strftime('%H:%M:%S %Y-%m-%d')
 
         image_urls = []
         if "images" in request.files:
@@ -382,7 +382,7 @@ def createnewpost():
             return redirect('/me')
         else:
             flash('exceeds the maximum word limit by '+str(-(POST_MAX - len(content)))+', sorry', 'danger')
-    return render_template('create_post.html', title='New Post', form=form)
+    return render_template('forms/edit_post.html', form=form, p=p)
 
 
 @app.route("/image/posts/<userid>/<imageuid>", methods=["GET"])
@@ -483,7 +483,7 @@ def settings():
         else:
             flash('Your "About Me" is too long.', 'warning')
 
-    return render_template('settings.html', title='Set About Me', form=form)
+    return render_template('forms/settings.html', title='Set About Me', form=form)
 
 
 
@@ -522,7 +522,7 @@ def request_reset_password():
         else:
             flash('There is no account with that email. You must register first.', 'warning')
         return redirect(url_for('login'))
-    return render_template('request_reset_password.html', title='Reset Password', form=form)
+    return render_template('forms/request_reset_password.html', title='Reset Password', form=form)
 
 
 @app.route("/reset_password/<token>", methods=['GET', 'POST'])
@@ -538,14 +538,14 @@ def reset_token(token):
         db.userdb.update_one({"_id": user.get_id()}, {"$set": {"password": user.password}})
         flash('Your password has been updated! You are now able to log in', 'success')
         return redirect(url_for('login'))
-    return render_template('reset_password.html', title='Reset Password', form=form)
+    return render_template('forms/reset_password.html', title='Reset Password', form=form)
 
 
 @app.route("/messaging/dashboard", methods=['GET', 'POST'])
 @login_required
 def messagingdashboard():
     conversations = Messages.get_conversations(current_user.username)
-    return render_template('mdashboard.html', conversations=conversations, User=User)
+    return render_template('messages/mdashboard.html', conversations=conversations, User=User)
 
 
 @app.route('/deletemsg')
@@ -569,7 +569,7 @@ def message_page(username):
     other_user = User.get_by_username(username)
     if not other_user:
         abort(404)
-    return render_template("message-page.html", username=username, other_user=other_user, User=User)
+    return render_template("messages/message-page.html", username=username, other_user=other_user, User=User)
 
 @app.route("/api/messages")
 @login_required
@@ -596,7 +596,7 @@ def get_messages():
     # Automatically mark messages as read when fetched via API
     Messages.mark_as_read(user2, current_user.username)
     # Update current user's last_seen
-    db.userdb.update_one({"_id": current_user._id}, {"$set": {"last_seen": datetime.utcnow()}})
+    db.userdb.update_one({"_id": current_user._id}, {"$set": {"last_seen": dt.utcnow()}})
 
     return jsonify([Messages(**m).json() for m in chats])
 
@@ -611,7 +611,7 @@ def send_message_v2():
         return jsonify({"success": False, "error": "Missing data"}), 400
 
     msg = Messages.send_message(current_user.username, recipient, content)
-    db.userdb.update_one({"_id": current_user._id}, {"$set": {"last_seen": datetime.utcnow()}})
+    db.userdb.update_one({"_id": current_user._id}, {"$set": {"last_seen": dt.utcnow()}})
     return jsonify({"success": True, "message": msg})
 
 @app.route('/save_theme', methods=['POST'])
@@ -766,7 +766,7 @@ def user_repos(username):
     if not user:
         abort(404)
     repos = Repository.find_by_owner(username)
-    return render_template("repos.html", user=user, repos=repos)
+    return render_template("repo/repos.html", user=user, repos=repos)
 
 
 @app.route("/repo/create", methods=["GET", "POST"])
@@ -777,12 +777,12 @@ def create_repo():
         name = secure_filename(form.name.data)
         if not name:
             flash("Invalid repository name.", "danger")
-            return render_template("create_repo.html", form=form)
+            return render_template("forms/create_repo.html", form=form)
 
         existing = Repository.get_by_owner_and_name(current_user.username, name)
         if existing:
             flash("You already have a repository with that name.", "danger")
-            return render_template("create_repo.html", form=form)
+            return render_template("forms/create_repo.html", form=form)
 
         repo_dir = os.path.join(REPOS_PATH, current_user.username, f"{name}.git")
         if os.path.exists(repo_dir):
@@ -800,9 +800,9 @@ def create_repo():
             return redirect(url_for("repo_view", username=current_user.username, reponame=name))
         except subprocess.CalledProcessError:
             flash("Error initializing repository.", "danger")
-            return render_template("create_repo.html", form=form)
+            return render_template("forms/create_repo.html", form=form)
 
-    return render_template("create_repo.html", form=form)
+    return render_template("forms/create_repo.html", form=form)
 
 
 @app.route("/repo/<username>/<reponame>")
@@ -848,7 +848,7 @@ def repo_view(username, reponame, ref=None, path=""):
 
     clone_url = f"{request.url_root.rstrip('/')}/git/{username}/{reponame}.git"
     return render_template(
-        "repo_view.html",
+        "repo/repo_view.html",
         repo=repo,
         clone_url=clone_url,
         entries=entries,
@@ -873,7 +873,7 @@ def repo_commits(username, reponame, ref):
     commits = get_commit_history(repo_dir, ref)
 
     return render_template(
-        "repo_commits.html",
+        "repo/repo_commits.html",
         repo=repo,
         ref=ref,
         commits=commits,
@@ -909,7 +909,7 @@ def repo_branches(username, reponame):
     except subprocess.CalledProcessError:
         branches = []
 
-    return render_template("repo_branches.html", repo=repo, branches=branches)
+    return render_template("repo/repo_branches.html", repo=repo, branches=branches)
 
 
 @app.route("/repo/<username>/<reponame>/blob/<ref>/<path:path>")
@@ -931,7 +931,7 @@ def repo_blob(username, reponame, ref, path):
         content = "[Binary content]"
 
     return render_template(
-        "repo_blob.html",
+        "repo/repo_blob.html",
         repo=repo,
         ref=ref,
         path=path,
@@ -1014,7 +1014,7 @@ def repo_edit(username, reponame, ref, path):
         content = "[Binary content]"
 
     return render_template(
-        "repo_edit.html",
+        "forms/repo_edit.html",
         repo=repo,
         ref=ref,
         path=path,
@@ -1132,7 +1132,7 @@ def get_latest_commit(repo_dir, ref):
             return {
                 "hash": lines[0],
                 "author": lines[1],
-                "date": datetime.fromtimestamp(int(lines[2])),
+                "date": dt.fromtimestamp(int(lines[2])),
                 "message": lines[3]
             }
     except:
@@ -1168,7 +1168,7 @@ def get_commit_history(repo_dir, ref):
                         "graph": graph,
                         "hash": commit_hash,
                         "author": parts[1],
-                        "date": datetime.fromtimestamp(int(parts[2])),
+                        "date": dt.fromtimestamp(int(parts[2])),
                         "message": parts[3]
                     })
                 else:
@@ -1287,11 +1287,12 @@ def edit_post(post_id):
     form = PostForm()
 
     if request.method == 'GET':
+        p = 1
         form.title.data = post.title
         form.content.data = post.content
 
     if form.validate_on_submit():
-        if len(form.content.data) <= POST_MAX:
+        if len(form.content.data) in range(POST_MIN,POST_MAX+1):
             db.postdb.update_one(
                 {"_id": post_id},
                 {
@@ -1304,7 +1305,165 @@ def edit_post(post_id):
             flash("Post updated successfully!", "success")
             return redirect(url_for('post_view', post_id=post_id))
         else:
-            flash("Post content too long", "warning")
+            if len(form.content.data) > POST_MAX:
+                flash("Post content too long", "warning")
+            else:
+                flash("Post content too short", "danger")
+    return render_template("forms/edit_post.html", form=form, p=p)
 
-    return render_template("edit_post.html", form=form)
 
+
+def generate_jwt(user_id):
+    payload = {
+        "user_id": user_id,
+        "exp": dt.utcnow() + timedelta(seconds=1)
+    }
+    return jwt.encode(payload, app.config['JWT_SECRET'], algorithm=app.config['JWT_ALGORITHM'])
+
+# Helper: Decorator to protect API routes
+def jwt_required(f):
+    @wraps(f)
+    def decorated_function(*args, **kwargs):
+        token = None
+        auth_header = request.headers.get('Authorization')
+        if auth_header and auth_header.startswith("Bearer "):
+            token = auth_header.split(" ")[1]
+        if not token:
+            return jsonify({"message": "Missing token"}), 401
+        try:
+            data = jwt.decode(token, app.config['JWT_SECRET'], algorithms=[app.config['JWT_ALGORITHM']])
+            user = User.get_by_id(data["user_id"])
+            if not user:
+                raise Exception("User not found")
+            g.current_user = user
+        except Exception as e:
+            return jsonify({"message": "Invalid or expired token", "error": str(e)}), 401
+        return f(*args, **kwargs)
+    return decorated_function
+
+@app.route("/api/login", methods=["POST"])
+def api_login():
+    data = request.get_json()
+    username = data.get("username")
+    password = data.get("password")
+    user = User.get_by_username(username)
+    if user and User.login_valid(username, password):
+        token = generate_jwt(user.get_id())
+        return jsonify({"success": True, "token": token, "username": username})
+    return jsonify({"success": False, "message": "Invalid credentials"}), 401
+
+@app.route('/api/me', methods=['GET'])
+@jwt_required
+def get_me():
+    user = g.current_user
+    return jsonify({
+        "username": user.username,
+        "email": user.email,
+        "aboutme": user.aboutme,
+        "avatar": User.avatar(user.username)
+    })
+
+@app.route('/api/user/<username>', methods=['GET'])
+@jwt_required
+def get_user_profile(username):
+    user = User.get_by_username(username)
+    if not user:
+        return jsonify({"error": "User not found"}), 404
+    return jsonify({
+        "username": user.username,
+        "email": user.email,
+        "aboutme": user.aboutme,
+        "followers": user.get_followers(),
+        "following": user.get_following(),
+        "avatar": User.avatar(user.username)
+    })
+
+@app.route('/api/user/<username>/follow', methods=['POST'])
+@jwt_required
+def api_toggle_follow(username):
+    if username == g.current_user.username:
+        return jsonify({"success": False, "message": "You cannot follow yourself!"}), 400
+
+    target = db.userdb.find_one({"username": username})
+    if not target:
+        return jsonify({"success": False, "message": "User not found!"}), 404
+
+    is_following = db.userdb.find_one({
+        "username": username,
+        "followers": g.current_user.username
+    })
+
+    if is_following:
+        db.userdb.update_one({"username": username}, {"$pull": {"followers": g.current_user.username}})
+        db.userdb.update_one({"username": g.current_user.username}, {"$pull": {"following": username}})
+        action = "unfollowed"
+    else:
+        db.userdb.update_one({"username": username}, {"$addToSet": {"followers": g.current_user.username}})
+        db.userdb.update_one({"username": g.current_user.username}, {"$addToSet": {"following": username}})
+        action = "followed"
+
+    updated_user = db.userdb.find_one({"username": username})
+    follower_count = len(updated_user.get("followers", []))
+
+    return jsonify({
+        "success": True,
+        "action": action,
+        "is_following": not is_following,
+        "follower_count": follower_count
+    })
+
+@app.route('/api/posts', methods=['GET'])
+@jwt_required
+def get_all_posts():
+    posts = db.postdb.find().sort("timestamp", -1)
+    result = []
+    for post in posts:
+        result.append({
+            "_id": post["_id"],
+            "username": post["username"],
+            "title": post["title"],
+            "content": post["content"],
+            "timestamp": post["timestamp"],
+            "likes": post.get("likes", 0),
+            "dislikes": post.get("dislikes", 0)
+        })
+    return jsonify(result)
+
+@app.route('/api/post/<post_id>/like', methods=['POST'])
+@jwt_required
+def api_like_post(post_id):
+    Post.liked(_id=post_id, userx=g.current_user.username)
+    post = db.postdb.find_one({"_id": post_id})
+    return jsonify({
+        "success": True,
+        "likes": post.get("likes", 0),
+        "dislikes": post.get("dislikes", 0)
+    })
+
+@app.route('/api/post/<post_id>/dislike', methods=['POST'])
+@jwt_required
+def api_dislike_post(post_id):
+    Post.disliked(_id=post_id, userx=g.current_user.username)
+    post = db.postdb.find_one({"_id": post_id})
+    return jsonify({
+        "success": True,
+        "likes": post.get("likes", 0),
+        "dislikes": post.get("dislikes", 0)
+    })
+
+@app.route('/api/messages/<username>', methods=['GET'])
+@jwt_required
+def get_chat_with_user(username):
+    messages = Messages.get_chat(g.current_user.username, username)
+    return jsonify([m.json() for m in messages])
+
+@app.route('/api/message/send', methods=['POST'])
+@jwt_required
+def send_message():
+    data = request.get_json()
+    receiver = data.get('receiver')
+    message = data.get('message')
+    if not receiver or not message:
+        return jsonify({"error": "Missing receiver or message"}), 400
+    msg = Messages.send_message(g.current_user.username, receiver, message)
+    return jsonify(msg)
