+++ templates/base.html
@@ -57,12 +57,6 @@
           <span class="d-lg-none">Upload</span>
         </a>
       </li>
-      <li class="nav-item">
-        <a class="nav-link d-flex align-items-center" id="theme-toggle">
-          <span class="material-symbols-outlined mr-2">brightness_6</span>
-          <span class="d-lg-none">Theme</span>
-        </a>
-      </li>
     </ul>
   </div>
 </nav>
