--- templates/create_post.html
@@ -1,32 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-<div class="container py-5">
-    <div class="row justify-content-center">
-        <div class="col-md-8 col-lg-6">
-            <div class="card shadow rounded-4">
-                <div class="card-body">
-                    <h2 class="text-center mb-4" style="font-family: 'Abril Fatface', cursive;">
-                        Make a New Post!
-                    </h2>
-
-                    <form action="" method="post" enctype="multipart/form-data">
-                        {{ form.hidden_tag() }}
-                        {% import "bootstrap/wtf.html" as wtf %}
-                        {{ wtf.form_field(form.title) }}
-                        {{ wtf.form_field(form.content) }}
-                        <div class="mb-3">
-                            {{ form.images.label(class="form-label") }}
-                            {{ form.images(class="form-control", multiple="multiple") }}
-                        </div>
-                        <button type="submit" class="btn btn-primary mt-3 d-flex align-items-center gap-1">
-                            <span class="material-symbols-outlined">arrow_upload_ready</span>
-                        </button>
-                    </form>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-{% endblock %}
-
