+++ templates/forms/upload_image.html
